requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
httpx>=0.27.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
